Refactor game types to remove model re-exports
