+++ game/types.ts
@@ -1,15 +1,6 @@
+import type { Company } from './models/Company';
 
-
-import type { RoomPurpose } from './roomPurposes';
-import { Company } from './models/Company';
-import { Structure } from './models/Structure';
-import { Room } from './models/Room';
-import { Zone } from './models/Zone';
-import { Planting } from './models/Planting';
-import { Plant } from './models/Plant';
-
-export { Company, Structure, Room, Zone, Planting, Plant };
-export type { RoomPurpose };
+export type { RoomPurpose } from './roomPurposes';
 
 export interface StructureBlueprint {
   id: string;
